Only validate style guide when compiling with Java 8
The checkstyle plugin that we're using requires Java 8.  This patch
updates the Maven config to skip style guide checking when using Java 7
(which we still want to support).

Our Travis config builds and tests under both 7, 8, and 9, so this will
still ensure that PRs can't be merged if they violate the style guide.